App.css removed. Unused img removed. All console logs and errors removed. Unused imports removed. Button folder cleaned up, all styles put in same file and unnecessary files removed. Unused button components removed. Toggle logic made DRY with one Toggle file.
