+++ src/components/Header.jsx
@@ -5,8 +5,7 @@ import { Link } from "react-router-dom";
 import { HamburgerMenu, StyledHeader } from "./Header.styles";
 import { GhostContainer, GhostWrapper } from "./GhostContainer.styles";
 import { ToggleDiv, MenuOverlay, StyledMenuLink, NavContainer, ToggleWrapper } from "./MenuOverlay.styles";
-import LanguageToggle from "./LanguageToggle";
-import ColorToggle from "./ColorToggle";
+import Toggle from "./Toggle";
 
 // Icons
 import GhostIcon from "./icons/Ghost";
@@ -97,12 +96,12 @@ export default function Header() {
 
                             <ToggleDiv>
                                 <span>{text[`toggleColor-${colorMode}`].toUpperCase()}</span>
-                                <ColorToggle />
+                                <Toggle type="color" />
                             </ToggleDiv>
 
                             <ToggleDiv>
                                 <span>{text.toggleLanguage.toUpperCase()}</span>
-                                <LanguageToggle />
+                                <Toggle type="language" />
                             </ToggleDiv>
                         </ToggleWrapper>
