--- src/components/buttons/ChooseProfileRow.styles.jsx
@@ -1,10 +1,0 @@
-import styled from "@emotion/styled";
-
-export const ChooseProfileRow = styled.div`
-    display: flex;
-    gap: 0.75rem;
-
-    > * {
-        flex: 1;
-    }
-`;
